refactor: change directory organization to be clean
