--- src/app/pages/Home/index.tsx
@@ -1,12 +1,0 @@
-import { useTheme } from "styled-components";
-import * as Styled from "./styles";
-
-const Home: React.FC = () => {
-  return (
-    <Styled.Container>
-      <div>teste</div>
-    </Styled.Container>
-  );
-};
-
-export default Home;
