--- src/app/pages/Home/styles.ts
@@ -1,5 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  background-color: ${({ theme }) => theme.colors.primary};
-`
