+++ src/assets/Corvete Z06.png
