+++ src/styles/globalStyles.ts
@@ -4,12 +4,15 @@ const GlobalStyles = createGlobalStyle`
   html,
   body {
     color: ${({ theme }) => theme.font.colors.pure};
-    background-color: ${({ theme }) => theme.colors.background};
     padding: 0;
     margin: 0;
-    font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
-      Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
+    font-family: Inter, sans-serif;
   }
+
+  h1 {
+    font-family: Archivo, sans-serif;
+  }
+
   a {
     color: inherit;
     text-decoration: none;
